Extract additional mixins from DerivaML to reduce class size
Create 5 new mixins to further modularize DerivaML:
- FeatureMixin: feature creation, lookup, deletion, and value listing
- DatasetMixin: dataset CRUD, element types, and bag download
- AssetMixin: asset creation and listing
- ExecutionMixin: execution creation, restoration, and status updates
- FileMixin: file management and dataset version operations

Fix circular imports by:
- Moving dataset type imports to TYPE_CHECKING in interfaces.py
- Using lazy imports in model/__init__.py for DatabaseModel/DerivaMLDatabase
- Removing @validate_call from methods with forward-reference return types
