+++ src/deriva_ml/core/base.py
@@ -14,62 +14,37 @@ Typical usage example:
 from __future__ import annotations  # noqa: I001
 
 # Standard library imports
-from collections import defaultdict
 import logging
 from datetime import datetime
-from itertools import chain
 from pathlib import Path
-from typing import Dict, Iterable, List, cast, TYPE_CHECKING, Any
+from typing import Dict, List, cast, TYPE_CHECKING, Any
 from typing_extensions import Self
-from urllib.parse import urlsplit
-
 
 # Third-party imports
 import requests
-from pydantic import ConfigDict, validate_call
 
 # Deriva imports
 from deriva.core import DEFAULT_SESSION_CONFIG, get_credential, urlquote
-
-import deriva.core.datapath as datapath
 from deriva.core.deriva_server import DerivaServer
 from deriva.core.ermrest_catalog import ErmrestCatalog, ErmrestSnapshot
-from deriva.core.ermrest_model import Key, Table
+from deriva.core.ermrest_model import Table
 from deriva.core.utils.core_utils import DEFAULT_LOGGER_OVERRIDES
 from deriva.core.utils.globus_auth_utils import GlobusNativeLogin
 
-from deriva_ml.core.exceptions import DerivaMLInvalidTerm
-from deriva_ml.core.definitions import (
-    ML_SCHEMA,
-    RID,
-    ColumnDefinition,
-    FileSpec,
-    MLVocab,
-    MLTable,
-    Status,
-    TableDefinition,
-)
+from deriva_ml.core.definitions import ML_SCHEMA, RID, Status, TableDefinition
 from deriva_ml.core.config import DerivaMLConfig
-from deriva_ml.core.exceptions import DerivaMLTableTypeError, DerivaMLException
-from deriva_ml.dataset.aux_classes import DatasetSpec, DatasetVersion
-from deriva_ml.dataset.dataset import Dataset
-from deriva_ml.dataset.dataset_bag import DatasetBag
-from deriva_ml.dataset.upload import asset_file_path, execution_rids
-from deriva_ml.dataset.history import iso_to_snap
-
-# Local imports
-from deriva_ml.execution.execution_configuration import ExecutionConfiguration
-from deriva_ml.execution.workflow import Workflow
-from deriva_ml.feature import Feature, FeatureRecord
-from deriva_ml.model.catalog import DerivaModel
-from deriva_ml.dataset.catalog_graph import CatalogGraph
+from deriva_ml.core.exceptions import DerivaMLException
 from deriva_ml.interfaces import DerivaMLCatalog
-from deriva_ml.schema.annotations import asset_annotation
 from deriva_ml.core.mixins import (
     VocabularyMixin,
     RidResolutionMixin,
     PathBuilderMixin,
     WorkflowMixin,
+    FeatureMixin,
+    DatasetMixin,
+    AssetMixin,
+    ExecutionMixin,
+    FileMixin,
 )
 
 # Optional debug imports
@@ -82,6 +57,7 @@ except ImportError:  # Graceful fallback if IceCream isn't installed.
 
 if TYPE_CHECKING:
     from deriva_ml.execution.execution import Execution
+    from deriva_ml.model.catalog import DerivaModel
 
 # Stop pycharm from complaining about undefined references.
 ml: DerivaML
@@ -92,6 +68,11 @@ class DerivaML(
     RidResolutionMixin,
     VocabularyMixin,
     WorkflowMixin,
+    FeatureMixin,
+    DatasetMixin,
+    AssetMixin,
+    ExecutionMixin,
+    FileMixin,
     DerivaMLCatalog,
 ):
     """Core class for machine learning operations on a Deriva catalog.
@@ -186,6 +167,8 @@ class DerivaML(
                 "Please check your credentials and make sure you have logged in."
             )
         self.catalog = server.connect_ermrest(catalog_id)
+        # Import here to avoid circular imports
+        from deriva_ml.model.catalog import DerivaModel
         self.model = DerivaModel(self.catalog.getCatalogModel(), domain_schema=domain_schema)
         self.use_minid = use_minid
 
@@ -518,821 +501,12 @@ class DerivaML(
         # Create table in domain schema using provided definition
         return self.model.schemas[self.domain_schema].create_table(table.model_dump())
 
-    @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
-    def create_asset(
-        self,
-        asset_name: str,
-        column_defs: Iterable[ColumnDefinition] | None = None,
-        fkey_defs: Iterable[ColumnDefinition] | None = None,
-        referenced_tables: Iterable[Table] | None = None,
-        comment: str = "",
-        schema: str | None = None,
-    ) -> Table:
-        """Creates an asset table.
 
-        Args:
-            asset_name: Name of the asset table.
-            column_defs: Iterable of ColumnDefinition objects to provide additional metadata for asset.
-            fkey_defs: Iterable of ForeignKeyDefinition objects to provide additional metadata for asset.
-            referenced_tables: Iterable of Table objects to which asset should provide foreign-key references to.
-            comment: Description of the asset table. (Default value = '')
-            schema: Schema in which to create the asset table.  Defaults to domain_schema.
+    # Methods moved to mixins:
+    # - create_asset, list_assets -> AssetMixin
+    # - create_feature, feature_record_class, delete_feature, lookup_feature, list_feature_values -> FeatureMixin
+    # - find_datasets, create_dataset, lookup_dataset, delete_dataset, list_dataset_element_types,
+    #   add_dataset_element_type, download_dataset_bag -> DatasetMixin
+    # - _update_status, create_execution, restore_execution -> ExecutionMixin
+    # - add_files, list_files, _bootstrap_versions, _synchronize_dataset_versions, _set_version_snapshot -> FileMixin
 
-        Returns:
-            Table object for the asset table.
-        """
-        # Initialize empty collections if None provided
-        column_defs = column_defs or []
-        fkey_defs = fkey_defs or []
-        referenced_tables = referenced_tables or []
-        schema = schema or self.domain_schema
-
-        # Add an asset type to vocabulary
-        self.add_term(MLVocab.asset_type, asset_name, description=f"A {asset_name} asset")
-
-        # Create the main asset table
-        asset_table = self.model.schemas[schema].create_table(
-            Table.define_asset(
-                schema,
-                asset_name,
-                column_defs=[c.model_dump() for c in column_defs],
-                fkey_defs=[fk.model_dump() for fk in fkey_defs],
-                comment=comment,
-            )
-        )
-
-        # Create an association table between asset and asset type
-        self.model.schemas[self.domain_schema].create_table(
-            Table.define_association(
-                [
-                    (asset_table.name, asset_table),
-                    ("Asset_Type", self.model.name_to_table("Asset_Type")),
-                ]
-            )
-        )
-
-        # Create references to other tables if specified
-        for t in referenced_tables:
-            asset_table.create_reference(self.model.name_to_table(t))
-
-        # Create an association table for tracking execution
-        atable = self.model.schemas[self.domain_schema].create_table(
-            Table.define_association(
-                [
-                    (asset_name, asset_table),
-                    (
-                        "Execution",
-                        self.model.schemas[self.ml_schema].tables["Execution"],
-                    ),
-                ]
-            )
-        )
-        atable.create_reference(self.model.name_to_table("Asset_Role"))
-
-        # Add asset annotations
-        asset_annotation(asset_table)
-        return asset_table
-
-    def list_assets(self, asset_table: Table | str) -> list[dict[str, Any]]:
-        """Lists contents of an asset table.
-
-        Returns a list of assets with their types for the specified asset table.
-
-        Args:
-            asset_table: Table or name of the asset table to list assets for.
-
-        Returns:
-            list[dict[str, Any]]: List of asset records, each containing:
-                - RID: Resource identifier
-                - Type: Asset type
-                - Metadata: Asset metadata
-
-        Raises:
-            DerivaMLException: If the table is not an asset table or doesn't exist.
-
-        Example:
-            >>> assets = ml.list_assets("tissue_types")
-            >>> for asset in assets:
-            ...     print(f"{asset['RID']}: {asset['Type']}")
-        """
-        # Validate and get asset table reference
-        asset_table = self.model.name_to_table(asset_table)
-        if not self.model.is_asset(asset_table):
-            raise DerivaMLException(f"Table {asset_table.name} is not an asset")
-
-        # Get path builders for asset and type tables
-        pb = self.pathBuilder()
-        asset_path = pb.schemas[asset_table.schema.name].tables[asset_table.name]
-        (
-            asset_type_table,
-            _,
-            _,
-        ) = self.model.find_association(asset_table, MLVocab.asset_type)
-        type_path = pb.schemas[asset_type_table.schema.name].tables[asset_type_table.name]
-
-        # Build a list of assets with their types
-        assets = []
-        for asset in asset_path.entities().fetch():
-            # Get associated asset types for each asset
-            asset_types = (
-                type_path.filter(type_path.columns[asset_table.name] == asset["RID"])
-                .attributes(type_path.Asset_Type)
-                .fetch()
-            )
-            # Combine asset data with its types
-            assets.append(
-                asset | {MLVocab.asset_type.value: [asset_type[MLVocab.asset_type.value] for asset_type in asset_types]}
-            )
-        return assets
-
-    @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
-    def create_feature(
-        self,
-        target_table: Table | str,
-        feature_name: str,
-        terms: list[Table | str] | None = None,
-        assets: list[Table | str] | None = None,
-        metadata: list[ColumnDefinition | Table | Key | str] | None = None,
-        optional: list[str] | None = None,
-        comment: str = "",
-    ) -> type[FeatureRecord]:
-        """Creates a new feature definition.
-
-        A feature represents a measurable property or characteristic that can be associated with records in the target
-        table. Features can include vocabulary terms, asset references, and additional metadata.
-
-        Args:
-            target_table: Table to associate the feature with (name or Table object).
-            feature_name: Unique name for the feature within the target table.
-            terms: Optional vocabulary tables/names whose terms can be used as feature values.
-            assets: Optional asset tables/names that can be referenced by this feature.
-            metadata: Optional columns, tables, or keys to include in a feature definition.
-            optional: Column names that are not required when creating feature instances.
-            comment: Description of the feature's purpose and usage.
-
-        Returns:
-            type[FeatureRecord]: Feature class for creating validated instances.
-
-        Raises:
-            DerivaMLException: If a feature definition is invalid or conflicts with existing features.
-
-        Examples:
-            Create a feature with confidence score:
-                >>> feature_class = ml.create_feature(
-                ...     target_table="samples",
-                ...     feature_name="expression_level",
-                ...     terms=["expression_values"],
-                ...     metadata=[ColumnDefinition(name="confidence", type=BuiltinTypes.float4)],
-                ...     comment="Gene expression measurement"
-                ... )
-        """
-        # Initialize empty collections if None provided
-        terms = terms or []
-        assets = assets or []
-        metadata = metadata or []
-        optional = optional or []
-
-        def normalize_metadata(m: Key | Table | ColumnDefinition | str) -> Key | Table | dict:
-            """Helper function to normalize metadata references."""
-            if isinstance(m, str):
-                return self.model.name_to_table(m)
-            elif isinstance(m, ColumnDefinition):
-                return m.model_dump()
-            else:
-                return m
-
-        # Validate asset and term tables
-        if not all(map(self.model.is_asset, assets)):
-            raise DerivaMLException("Invalid create_feature asset table.")
-        if not all(map(self.model.is_vocabulary, terms)):
-            raise DerivaMLException("Invalid create_feature asset table.")
-
-        # Get references to required tables
-        target_table = self.model.name_to_table(target_table)
-        execution = self.model.schemas[self.ml_schema].tables["Execution"]
-        feature_name_table = self.model.schemas[self.ml_schema].tables["Feature_Name"]
-
-        # Add feature name to vocabulary
-        feature_name_term = self.add_term("Feature_Name", feature_name, description=comment)
-        atable_name = f"Execution_{target_table.name}_{feature_name_term.name}"
-        # Create an association table implementing the feature
-        atable = self.model.schemas[self.domain_schema].create_table(
-            target_table.define_association(
-                table_name=atable_name,
-                associates=[execution, target_table, feature_name_table],
-                metadata=[normalize_metadata(m) for m in chain(assets, terms, metadata)],
-                comment=comment,
-            )
-        )
-        # Configure optional columns and default feature name
-        for c in optional:
-            atable.columns[c].alter(nullok=True)
-        atable.columns["Feature_Name"].alter(default=feature_name_term.name)
-
-        # Return feature record class for creating instances
-        return self.feature_record_class(target_table, feature_name)
-
-    def feature_record_class(self, table: str | Table, feature_name: str) -> type[FeatureRecord]:
-        """Returns a pydantic model class for feature records.
-
-        Creates a typed interface for creating new instances of the specified feature. The returned class includes
-        validation and type checking based on the feature's definition.
-
-        Args:
-            table: The table containing the feature, either as name or Table object.
-            feature_name: Name of the feature to create a record class for.
-
-        Returns:
-            type[FeatureRecord]: A pydantic model class for creating validated feature records.
-
-        Raises:
-            DerivaMLException: If the feature doesn't exist or the table is invalid.
-
-        Example:
-            >>> ExpressionFeature = ml.feature_record_class("samples", "expression_level")
-            >>> feature = ExpressionFeature(value="high", confidence=0.95)
-        """
-        # Look up a feature and return its record class
-        return self.lookup_feature(table, feature_name).feature_record_class()
-
-    def delete_feature(self, table: Table | str, feature_name: str) -> bool:
-        """Removes a feature definition and its data.
-
-        Deletes the feature and its implementation table from the catalog. This operation cannot be undone and
-        will remove all feature values associated with this feature.
-
-        Args:
-            table: The table containing the feature, either as name or Table object.
-            feature_name: Name of the feature to delete.
-
-        Returns:
-            bool: True if the feature was successfully deleted, False if it didn't exist.
-
-        Raises:
-            DerivaMLException: If deletion fails due to constraints or permissions.
-
-        Example:
-            >>> success = ml.delete_feature("samples", "obsolete_feature")
-            >>> print("Deleted" if success else "Not found")
-        """
-        # Get table reference and find feature
-        table = self.model.name_to_table(table)
-        try:
-            # Find and delete the feature's implementation table
-            feature = next(f for f in self.model.find_features(table) if f.feature_name == feature_name)
-            feature.feature_table.drop()
-            return True
-        except StopIteration:
-            return False
-
-    def lookup_feature(self, table: str | Table, feature_name: str) -> Feature:
-        """Retrieves a Feature object.
-
-        Looks up and returns a Feature object that provides an interface to work with an existing feature
-        definition in the catalog.
-
-        Args:
-            table: The table containing the feature, either as name or Table object.
-            feature_name: Name of the feature to look up.
-
-        Returns:
-            Feature: An object representing the feature and its implementation.
-
-        Raises:
-            DerivaMLException: If the feature doesn't exist in the specified table.
-
-        Example:
-            >>> feature = ml.lookup_feature("samples", "expression_level")
-            >>> print(feature.feature_name)
-            'expression_level'
-        """
-        return self.model.lookup_feature(table, feature_name)
-
-    @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
-    def list_feature_values(self, table: Table | str, feature_name: str) -> datapath._ResultSet:
-        """Retrieves all values for a feature.
-
-        Returns all instances of the specified feature that have been created, including their associated
-        metadata and references.
-
-        Args:
-            table: The table containing the feature, either as name or Table object.
-            feature_name: Name of the feature to retrieve values for.
-
-        Returns:
-            datapath._ResultSet: A result set containing all feature values and their metadata.
-
-        Raises:
-            DerivaMLException: If the feature doesn't exist or cannot be accessed.
-
-        Example:
-            >>> values = ml.list_feature_values("samples", "expression_level")
-            >>> for value in values:
-            ...     print(f"Sample {value['RID']}: {value['value']}")
-        """
-        # Get table and feature references
-        table = self.model.name_to_table(table)
-        feature = self.lookup_feature(table, feature_name)
-
-        # Build and execute query for feature values
-        pb = self.pathBuilder()
-        return pb.schemas[feature.feature_table.schema.name].tables[feature.feature_table.name].entities().fetch()
-
-    # add_term, lookup_term, list_vocabulary_terms moved to VocabularyMixin
-    # get_table_as_dataframe, get_table_as_dict moved to PathBuilderMixin
-
-    def find_datasets(self, deleted: bool = False) -> Iterable[Dataset]:
-        """Returns a list of currently available datasets.
-
-        Arguments:
-            deleted: If True, included the datasets that have been deleted.
-
-        Returns:
-             list of currently available datasets.
-        """
-        # Get datapath to all the tables we will need: Dataset, DatasetType and the association table.
-        pb = self.pathBuilder()
-        dataset_path = pb.schemas[self._dataset_table.schema.name].tables[self._dataset_table.name]
-        associations = list(self.model.schemas[self.ml_schema].tables[MLVocab.dataset_type].find_associations())
-        atable = associations[0].name if associations else None
-        ml_path = pb.schemas[self.ml_schema]
-        atable_path = ml_path.tables[atable]
-
-        if deleted:
-            filtered_path = dataset_path
-        else:
-            filtered_path = dataset_path.filter(
-                (dataset_path.Deleted == False) | (dataset_path.Deleted == None)  # noqa: E711, E712
-            )
-
-        # Get a list of all the dataset_type values associated with this dataset_table.
-        datasets = []
-        for dataset in filtered_path.entities().fetch():
-            ds_types = (
-                atable_path.filter(atable_path.Dataset == dataset["RID"]).attributes(atable_path.Dataset_Type).fetch()
-            )
-            datasets.append(
-                Dataset(
-                    self,
-                    dataset_rid=dataset["RID"],
-                    description=dataset["Description"],
-                    dataset_types=[ds[MLVocab.dataset_type] for ds in ds_types],
-                )
-            )
-        return datasets
-
-    def create_dataset(
-        self,
-        version: DatasetVersion | str | None = None,
-        execution_rid: RID | None = None,
-        description: str = "",
-        dataset_types: list[str] | None = None,
-    ) -> Dataset:
-        return Dataset.create_dataset(
-            self, execution_rid=execution_rid, version=version, description=description, dataset_types=dataset_types
-        )
-
-    def lookup_dataset(self, dataset: RID | DatasetSpec, deleted: bool = False
-    ) -> Dataset:
-        """Looks up a dataset by RID or DatasetSpec.
-
-        Arguments:
-            dataset: Dataset RID or DatasetSpec to look up.
-            version specified in the DatasetSpec.
-            deleted: If True, included the datasets that have been deleted.
-        Returns a Dataset object for the specified dataset RID or DatasetSpec.
-
-        """
-        if isinstance(dataset, DatasetSpec):
-            dataset_rid = dataset.rid
-        else:
-            dataset_rid = dataset
-
-        try:
-            return [ds for ds in self.find_datasets(deleted=deleted) if ds.dataset_rid == dataset_rid][0]
-        except IndexError:
-            raise DerivaMLException(f"Dataset {dataset_rid} not found.")
-
-    @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
-    def delete_dataset(self, dataset: Dataset, recurse: bool = False) -> None:
-        """Delete a dataset_table from the catalog.
-
-        Args:
-            dataset: RID of the dataset_table to delete.
-            recurse: If True, delete the dataset_table along with any nested datasets. (Default value = False)
-        """
-        # Get association table entries for this dataset_table
-        # Delete association table entries
-        dataset_rid = dataset.dataset_rid
-        if not self.model.is_dataset_rid(dataset.dataset_rid):
-            raise DerivaMLException("Dataset_rid is not a dataset.")
-
-        if parents := dataset.list_dataset_parents():
-            raise DerivaMLException(f'Dataset "{dataset}" is in a nested dataset: {parents}.')
-
-        pb = self.pathBuilder()
-        dataset_path = pb.schemas[self._dataset_table.schema.name].tables[self._dataset_table.name]
-
-        rid_list = [dataset_rid] + (dataset.list_dataset_children() if recurse else [])
-        dataset_path.update([{"RID": r, "Deleted": True} for r in rid_list])
-
-    def list_dataset_element_types(self) -> Iterable[Table]:
-        """List the types of entities that can be added to a dataset_table.
-
-        Returns:
-          :return: An iterable of Table objects that can be included as an element of a dataset_table.
-        """
-
-        def domain_table(table: Table) -> bool:
-            return table.schema.name == self.domain_schema or table.name == self._dataset_table.name
-
-        return [t for a in self._dataset_table.find_associations() if domain_table(t := a.other_fkeys.pop().pk_table)]
-
-    @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
-    def add_dataset_element_type(self, element: str | Table) -> Table:
-        """A dataset_table is a heterogeneous collection of objects, each of which comes from a different table. This
-        routine makes it possible to add objects from the specified table to a dataset_table.
-
-        Args:
-            element: Name of the table or table object that is to be added to the dataset_table.
-
-        Returns:
-            The table object that was added to the dataset_table.
-        """
-        # Add table to map
-        element_table = self.model.name_to_table(element)
-        atable_def = Table.define_association([self._dataset_table, element_table])
-        try:
-            table = self.model.schemas[self.model.domain_schema].create_table(atable_def)
-        except ValueError as e:
-            if "already exists" in str(e):
-                table = self.model.name_to_table(atable_def["table_name"])
-            else:
-                raise e
-
-        # self.model = self.catalog.getCatalogModel()
-        annotations = CatalogGraph(self, self.use_minid).generate_dataset_download_annotations()
-        self._dataset_table.annotations.update(annotations)
-        self.model.model.apply()
-        return table
-
-    @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
-    def download_dataset_bag(
-        self,
-        dataset: DatasetSpec,
-    ) -> DatasetBag:
-        """Downloads a dataset to the local filesystem and creates a MINID if needed.
-
-        Downloads a dataset specified by DatasetSpec to the local filesystem. If the dataset doesn't have
-        a MINID (Minimal Viable Identifier), one will be created. The dataset can optionally be associated
-        with an execution record.
-
-        Args:
-            dataset: Specification of the dataset to download, including version and materialization options.
-            execution_rid: Optional execution RID to associate the download with.
-
-        Returns:
-            DatasetBag: Object containing:
-                - path: Local filesystem path to downloaded dataset
-                - rid: Dataset's Resource Identifier
-                - minid: Dataset's Minimal Viable Identifier
-
-        Examples:
-            Download with default options:
-                >>> spec = DatasetSpec(rid="1-abc123")
-                >>> bag = ml.download_dataset_bag(dataset=spec)
-                >>> print(f"Downloaded to {bag.path}")
-
-            Download with execution tracking:
-                >>> bag = ml.download_dataset_bag(
-                ...     dataset=DatasetSpec(rid="1-abc123", materialize=True),
-                ...     execution_rid="1-xyz789"
-                ... )
-        """
-        if not self.model.is_dataset_rid(dataset.rid):
-            raise DerivaMLTableTypeError("Dataset", dataset.rid)
-        ds = self.lookup_dataset(dataset)
-        return ds.download_dataset_bag(
-            version=dataset.version,
-            materialize=dataset.materialize,
-            use_minid=self.use_minid,
-        )
-
-    def _update_status(self, new_status: Status, status_detail: str, execution_rid: RID) -> None:
-        """Update the status of an execution in the catalog.
-
-        Args:
-            new_status: New status.
-            status_detail: Details of the status.
-            execution_rid: Resource Identifier (RID) of the execution.
-            new_status: Status:
-            status_detail: str:
-             execution_rid: RID:
-
-        Returns:
-
-        """
-        self.status = new_status.value
-        self.pathBuilder().schemas[self.ml_schema].Execution.update(
-            [
-                {
-                    "RID": execution_rid,
-                    "Status": self.status,
-                    "Status_Detail": status_detail,
-                }
-            ]
-        )
-
-    @validate_call(config=ConfigDict(arbitrary_types_allowed=True))
-    def add_files(
-        self,
-        files: Iterable[FileSpec],
-        dataset_types: str | list[str] | None = None,
-        description: str = "",
-        execution_rid: RID | None = None,
-    ) -> Dataset:
-        """Adds files to the catalog with their metadata.
-
-        Registers files in the catalog along with their metadata (MD5, length, URL) and associates them with
-        specified file types. Optionally links files to an execution record.
-
-        Args:
-            files: File specifications containing MD5 checksum, length, and URL.
-            dataset_types: One or more dataset type terms from File_Type vocabulary.
-            description: Description of the files.
-            execution_rid: Optional execution RID to associate files with.
-
-        Returns:
-            RID: Resource of dataset that represents the newly added files.
-
-        Raises:
-            DerivaMLException: If file_types are invalid or execution_rid is not an execution record.
-
-        Examples:
-            Add a single file type:
-                >>> files = [FileSpec(url="path/to/file.txt", md5="abc123", length=1000)]
-                >>> rids = ml.add_files(files, file_types="text")
-
-            Add multiple file types:
-                >>> rids = ml.add_files(
-                ...     files=[FileSpec(url="image.png", md5="def456", length=2000)],
-                ...     file_types=["image", "png"],
-                ...     execution_rid="1-xyz789"
-                ... )
-        """
-        if execution_rid and self.resolve_rid(execution_rid).table.name != "Execution":
-            raise DerivaMLTableTypeError("Execution", execution_rid)
-
-        filespec_list = list(files)
-
-        # Get a list of all defined file types and their synonyms.
-        defined_types = set(
-            chain.from_iterable([[t.name] + t.synonyms for t in self.list_vocabulary_terms(MLVocab.asset_type)])
-        )
-
-        # Get a list of al of the file types used in the filespec_list
-        spec_types = set(chain.from_iterable(filespec.file_types for filespec in filespec_list))
-
-        # Now make sure that all of the file types and dataset_types in the spec list are defined.
-        if spec_types - defined_types:
-            raise DerivaMLInvalidTerm(MLVocab.asset_type.name, f"{spec_types - defined_types}")
-
-        # Normalize dataset_types, make sure FIle type is included.
-        if isinstance(dataset_types, list):
-            dataset_types = ["File"] + dataset_types if "File" not in dataset_types else dataset_types
-        else:
-            dataset_types = ["File", dataset_types] if dataset_types else ["File"]
-        for ds_type in dataset_types:
-            self.lookup_term(MLVocab.dataset_type, ds_type)
-
-        # Add files to the file table, and collect up the resulting entries by directory name.
-        pb = self.pathBuilder()
-        file_records = list(
-            pb.schemas[self.ml_schema].tables["File"].insert([f.model_dump(by_alias=True) for f in filespec_list])
-        )
-
-        # Get the name of the association table between file_table and file_type and add file_type records
-        atable = self.model.find_association(MLTable.file, MLVocab.asset_type)[0].name
-        # Need to get a link between file record and file_types.
-        type_map = {
-            file_spec.md5: file_spec.file_types + ([] if "File" in file_spec.file_types else [])
-            for file_spec in filespec_list
-        }
-        file_type_records = [
-            {MLVocab.asset_type.value: file_type, "File": file_record["RID"]}
-            for file_record in file_records
-            for file_type in type_map[file_record["MD5"]]
-        ]
-        pb.schemas[self.ml_schema].tables[atable].insert(file_type_records)
-
-        if execution_rid:
-            # Get the name of the association table between file_table and execution.
-            pb.schemas[self.ml_schema].File_Execution.insert(
-                [
-                    {"File": file_record["RID"], "Execution": execution_rid, "Asset_Role": "Output"}
-                    for file_record in file_records
-                ]
-            )
-
-        # Now create datasets to capture the original directory structure of the files.
-        dir_rid_map = defaultdict(list)
-        for e in file_records:
-            dir_rid_map[Path(urlsplit(e["URL"]).path).parent].append(e["RID"])
-
-        nested_datasets = []
-        path_length = 0
-        dataset = None
-        # Start with the longest path so we get subdirectories first.
-        for p, rids in sorted(dir_rid_map.items(), key=lambda kv: len(kv[0].parts), reverse=True):
-            dataset = Dataset.create_dataset(
-                self, dataset_types=dataset_types, execution_rid=execution_rid, description=description
-            )
-            members = rids
-            if len(p.parts) < path_length:
-                # Going up one level in directory, so Create nested dataset
-                members = [m.dataset_rid for m in nested_datasets] + rids
-                nested_datasets = []
-            dataset.add_dataset_members(members=members, execution_rid=execution_rid)
-            nested_datasets.append(dataset)
-            path_length = len(p.parts)
-
-        return dataset
-
-    def _bootstrap_versions(self) -> None:
-        datasets = [ds.dataset_rid for ds in self.find_datasets()]
-        ds_version = [
-            {
-                "Dataset": d,
-                "Version": "0.1.0",
-                "Description": "Dataset at the time of conversion to versioned datasets",
-            }
-            for d in datasets
-        ]
-        schema_path = self.pathBuilder().schemas[self.ml_schema]
-        version_path = schema_path.tables["Dataset_Version"]
-        dataset_path = schema_path.tables["Dataset"]
-        history = list(version_path.insert(ds_version))
-        dataset_versions = [{"RID": h["Dataset"], "Version": h["Version"]} for h in history]
-        dataset_path.update(dataset_versions)
-
-    def _synchronize_dataset_versions(self) -> None:
-        schema_path = self.pathBuilder().schemas[self.ml_schema]
-        dataset_version_path = schema_path.tables["Dataset_Version"]
-        # Get the maximum version number for each dataset.
-        versions = {}
-        for v in dataset_version_path.entities().fetch():
-            if v["Version"] > versions.get("Dataset", DatasetVersion(0, 0, 0)):
-                versions[v["Dataset"]] = v
-        dataset_path = schema_path.tables["Dataset"]
-        dataset_path.update([{"RID": dataset, "Version": version["RID"]} for dataset, version in versions.items()])
-
-    def _set_version_snapshot(self) -> None:
-        """Update the Snapshot column of the Dataset_Version table to the correct time."""
-        dataset_version_path = self.pathBuilder().schemas[self.model.ml_schema].tables["Dataset_Version"]
-        versions = dataset_version_path.entities().fetch()
-        dataset_version_path.update(
-            [{"RID": h["RID"], "Snapshot": iso_to_snap(h["RCT"])} for h in versions if not h["Snapshot"]]
-        )
-
-    def list_files(self, file_types: list[str] | None = None) -> list[dict[str, Any]]:
-        """Lists files in the catalog with their metadata.
-
-        Returns a list of files with their metadata including URL, MD5 hash, length, description,
-        and associated file types. Files can be optionally filtered by type.
-
-        Args:
-            file_types: Filter results to only include these file types.
-
-        Returns:
-            list[dict[str, Any]]: List of file records, each containing:
-                - RID: Resource identifier
-                - URL: File location
-                - MD5: File hash
-                - Length: File size
-                - Description: File description
-                - File_Types: List of associated file types
-
-        Examples:
-            List all files:
-                >>> files = ml.list_files()
-                >>> for f in files:
-                ...     print(f"{f['RID']}: {f['URL']}")
-
-            Filter by file type:
-                >>> image_files = ml.list_files(["image", "png"])
-        """
-
-        asset_type_atable, file_fk, asset_type_fk = self.model.find_association("File", "Asset_Type")
-        ml_path = self.pathBuilder().schemas[self.ml_schema]
-        file = ml_path.File
-        asset_type = ml_path.tables[asset_type_atable.name]
-
-        path = file.path
-        path = path.link(asset_type.alias("AT"), on=file.RID == asset_type.columns[file_fk], join_type="left")
-        if file_types:
-            path = path.filter(asset_type.columns[asset_type_fk] == datapath.Any(*file_types))
-        path = path.attributes(
-            path.File.RID,
-            path.File.URL,
-            path.File.MD5,
-            path.File.Length,
-            path.File.Description,
-            path.AT.columns[asset_type_fk],
-        )
-
-        file_map = {}
-        for f in path.fetch():
-            entry = file_map.setdefault(f["RID"], {**f, "File_Types": []})
-            if ft := f.get("Asset_Type"):  # assign-and-test in one go
-                entry["File_Types"].append(ft)
-
-        # Now get rid of the File_Type key and return the result
-        return [(f, f.pop("Asset_Type"))[0] for f in file_map.values()]
-
-    # list_workflows, add_workflow, lookup_workflow, create_workflow moved to WorkflowMixin
-
-    def create_execution(
-        self, configuration: ExecutionConfiguration, workflow: Workflow | RID | None = None, dry_run: bool = False
-    ) -> "Execution":
-        """Creates an execution environment.
-
-        Given an execution configuration, initialize the local compute environment to prepare for executing an
-        ML or analytic routine.  This routine has a number of side effects.
-
-        1. The datasets specified in the configuration are downloaded and placed in the cache-dir. If a version is
-        not specified in the configuration, then a new minor version number is created for the dataset and downloaded.
-
-        2. If any execution assets are provided in the configuration, they are downloaded
-        and placed in the working directory.
-
-
-        Args:
-            configuration: ExecutionConfiguration:
-            workflow: Workflow object representing the workflow to execute if not present in the ExecutionConfiguration.
-            dry_run: Do not create an execution record or upload results.
-
-        Returns:
-            An execution object.
-        """
-        # Import here to avoid circular dependency
-        from deriva_ml.execution.execution import Execution
-
-        # Create and store an execution instance
-        self._execution = Execution(configuration, self, workflow=workflow, dry_run=dry_run)
-        return self._execution
-
-    def restore_execution(self, execution_rid: RID | None = None) -> Execution:
-        """Restores a previous execution.
-
-        Given an execution RID, retrieves the execution configuration and restores the local compute environment.
-        This routine has a number of side effects.
-
-        1. The datasets specified in the configuration are downloaded and placed in the cache-dir. If a version is
-        not specified in the configuration, then a new minor version number is created for the dataset and downloaded.
-
-        2. If any execution assets are provided in the configuration, they are downloaded and placed
-        in the working directory.
-
-        Args:
-            execution_rid: Resource Identifier (RID) of the execution to restore.
-
-        Returns:
-            Execution: An execution object representing the restored execution environment.
-
-        Raises:
-            DerivaMLException: If execution_rid is not valid or execution cannot be restored.
-
-        Example:
-            >>> execution = ml.restore_execution("1-abc123")
-        """
-        # Import here to avoid circular dependency
-        from deriva_ml.execution.execution import Execution
-
-        # If no RID provided, try to find single execution in working directory
-        if not execution_rid:
-            e_rids = execution_rids(self.working_dir)
-            if len(e_rids) != 1:
-                raise DerivaMLException(f"Multiple execution RIDs were found {e_rids}.")
-            execution_rid = e_rids[0]
-
-        # Try to load configuration from a file
-        cfile = asset_file_path(
-            prefix=self.working_dir,
-            exec_rid=execution_rid,
-            file_name="configuration.json",
-            asset_table=self.model.name_to_table("Execution_Metadata"),
-            metadata={},
-        )
-
-        # Load configuration from a file or create from an execution record
-        if cfile.exists():
-            configuration = ExecutionConfiguration.load_configuration(cfile)
-        else:
-            execution = self.retrieve_rid(execution_rid)
-            configuration = ExecutionConfiguration(
-                workflow=execution["Workflow"],
-                description=execution["Description"],
-            )
-
-        # Create and return an execution instance
-        return Execution(configuration, self, reload=execution_rid)
