+++ src/deriva_ml/core/mixins/__init__.py
@@ -9,16 +9,31 @@ Mixins:
     RidResolutionMixin: RID resolution and retrieval
     PathBuilderMixin: Path building and table access utilities
     WorkflowMixin: Workflow management (add, lookup, list, create)
+    FeatureMixin: Feature management (create, lookup, delete, list values)
+    DatasetMixin: Dataset management (find, create, lookup, delete)
+    AssetMixin: Asset management (create, list assets)
+    ExecutionMixin: Execution management (create, restore, update status)
+    FileMixin: File management (add, list files)
 """
 
 from deriva_ml.core.mixins.vocabulary import VocabularyMixin
 from deriva_ml.core.mixins.rid_resolution import RidResolutionMixin
 from deriva_ml.core.mixins.path_builder import PathBuilderMixin
 from deriva_ml.core.mixins.workflow import WorkflowMixin
+from deriva_ml.core.mixins.feature import FeatureMixin
+from deriva_ml.core.mixins.dataset import DatasetMixin
+from deriva_ml.core.mixins.asset import AssetMixin
+from deriva_ml.core.mixins.execution import ExecutionMixin
+from deriva_ml.core.mixins.file import FileMixin
 
 __all__ = [
     "VocabularyMixin",
     "RidResolutionMixin",
     "PathBuilderMixin",
     "WorkflowMixin",
+    "FeatureMixin",
+    "DatasetMixin",
+    "AssetMixin",
+    "ExecutionMixin",
+    "FileMixin",
 ]
